requirements.txt
streamlit==1.31.0
pandas==2.2.0
openpyxl==3.1.2
folium==0.15.1
streamlit-folium==0.18.0
plotly==5.18.0
arabic-reshaper==3.0.0
python-bidi==0.4.2
